refactor: Add folder structure and rename files
