+++ src/components/App/App.component.jsx
@@ -1,7 +1,6 @@
-/* eslint-disable react/jsx-filename-extension */
 import React from 'react';
 import logo from './logo.svg';
-import './App.css';
+import './App.styles.css';
 
 function App() {
   return (
