+++ src/components/App/App.test.jsx
@@ -1,7 +1,6 @@
-/* eslint-disable react/jsx-filename-extension */
 import React from 'react';
 import { render } from '@testing-library/react';
-import App from './App';
+import App from './App.component';
 
 test('renders learn react link', () => {
   const { getByText } = render(<App />);
